+++ tests/qrcode/test_fake_qrcode_segno.py
@@ -1,16 +1,12 @@
 # -*- coding: utf-8 -*-
 #
-# Copyright (c) 2016 -- Lars Heuer - Semagia <http://www.semagia.com/>.
+# Copyright (c) 2016 - 2017 -- Lars Heuer - Semagia <http://www.semagia.com/>.
 # All rights reserved.
 #
 # License: BSD License
 #
 """\
 Tests if install_as_qrcode works
-
-:author:       Lars Heuer (heuer[at]semagia.com)
-:organization: Semagia - http://www.semagia.com/
-:license:      BSD License
 """
 from __future__ import absolute_import, unicode_literals
 import segno_mimos
@@ -18,7 +14,7 @@ import pytest
 import io
 
 
-def test_svgfragment():
+def test_install_segno_as_qrcode():
     segno_mimos.install_as_qrcode()
     import qrcode
     import qrcode.image.svg